# zoho-middleware/openapi.yaml
openapi: 3.0.3
info:
  title: Steins & Vines Middleware API
  description: |
    Backend middleware for the Steins & Vines fermentation supply store.
    Bridges the frontend (GitHub Pages) with Zoho Inventory, Zoho Books,
    Zoho Bookings, and Global Payments.

    **Authentication model:**
    - Most `GET` endpoints are public (no key required).
    - Most `POST`/`PUT` endpoints require `X-Api-Key: <API_SECRET_KEY>`.
    - `POST /api/checkout` is public — protected by reCAPTCHA + rate limiting instead.
    - `POST /product-requests` and `POST /api/contact` are fully public.
    - Auth routes (`/auth/*`) are pre-auth and never require a key.

    **CORS:** Requests must originate from `steinsandvines.ca`,
    `staging.steinsandvines.ca`, `localhost:3001`, or `localhost:8080`.

    **Rate limits:** `/api/*` endpoints are limited to 60 req/min per IP.
    Payment endpoints (`/api/checkout`, `/api/payment/*`, `/api/pos/sale`,
    `/api/kiosk/sale`) share a stricter 10 req/min limiter.
  version: "1.0.0"
  contact:
    name: Steins & Vines IT
    email: hello@steinsandvines.ca
servers:
  - url: https://steinsandvines-middleware.railway.app
    description: Production (Railway)
  - url: http://localhost:3001
    description: Local development
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Must equal the `API_SECRET_KEY` Railway env var.
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Cart is empty"
...
paths:
...
  /api/pos/collect:
    post:
      tags: [pos]
      summary: Send SO balance to Helcim terminal for collection
      description: >
        Called by Zoho Inventory custom button (Deluge invokeUrl).
        Fetches the Sales Order balance and pushes it to the Helcim
        Smart Terminal. Returns 202 immediately; webhook handles
        the payment result and updates Zoho.
      operationId: postPosCollect
      security:
        - ApiKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [salesorder_id]
              properties:
                salesorder_id:
                  type: string
                  description: Zoho Sales Order ID
      responses:
        "202":
          description: Payment sent to terminal.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  salesorder_number:
                    type: string
                  amount:
                    type: number
                  status:
                    type: string
                    enum: [pending]
        "400":
          description: Invalid input, zero balance, or void/closed order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Sales Order not found in Zoho.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Payment already pending for this order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "503":
          description: Helcim terminal not configured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "502":
          description: Terminal or Zoho API error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...